requirements.txt
dotenv==0.9.9
numpy==2.2.5
pandas==2.2.3
psycopg2-binary==2.9.10
pymongo==4.12.0

